Backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cloudinary==1.44.0
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
dj-database-url==3.1.2
Django==5.2.6
django-cors-headers==4.9.0
django-dotenv==1.4.2
django-filter==25.1
django-jazzmin==3.0.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dotenv==0.9.9
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.6.4
packaging==25.0
pillow==12.1.0
psycopg2==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
PySocks==1.7.1
python-dotenv==1.2.2
python3-openid==3.2.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
